Allow entity to lack a type if we can find it by coreference
